Refactor task worker and utils
