+++ app/views/views.py
@@ -8,7 +8,8 @@ from django.views import View
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.csrf import csrf_exempt
 from django.shortcuts import render, get_object_or_404
-from app.tasks import parser, generate_pdf
+from app.worker.parser import parser
+from app.worker.generate_pdf import generate_pdf
 from app.models import Donor, Donation, Item
 import csv
 import simplejson as json
