+++ app/worker/__init__.py
@@ -1,0 +1,3 @@
+'''
+Module for tasks to be sent on task queue 
+'''
